requirements.txt
certifi==2020.11.8
chardet==3.0.4
idna==2.10
pushbullet.py==0.12.0
python-magic==0.4.18
PyYAML==5.3.1
requests==2.25.0
six==1.15.0
urllib3==1.26.2
websocket-client==0.57.0
